file: class files move to cursor package
