+++ src/main/java/com/mide/gangsaeng/common/cursor/Cursor.java
@@ -1,8 +1,7 @@
-package com.mide.gangsaeng.common;
+package com.mide.gangsaeng.common.cursor;
 
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
 @Builder
